requirements.txt
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
decorator==5.0.9
emoji==1.2.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
Pillow==8.2.0
pyparsing==2.4.7
pyTelegramBotAPI==3.7.9
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
self==2020.12.3
six==1.16.0
speedtest-cli==2.1.3
urllib3==1.26.5

discord~=1.0.1